fix code according to advices
