+++ app/src/androidTest/java/org/mozilla/focus/autobot/Autobot.kt
@@ -11,8 +11,8 @@ import android.support.test.espresso.IdlingRegistry
 import android.support.test.espresso.IdlingResource
 import android.support.test.espresso.action.ViewActions
 import android.support.test.espresso.action.ViewActions.click
-import android.support.test.espresso.action.ViewActions.replaceText
 import android.support.test.espresso.action.ViewActions.pressImeActionButton
+import android.support.test.espresso.action.ViewActions.replaceText
 import android.support.test.espresso.assertion.ViewAssertions
 import android.support.test.espresso.assertion.ViewAssertions.matches
 import android.support.test.espresso.contrib.RecyclerViewActions
@@ -20,10 +20,6 @@ import android.support.test.espresso.matcher.ViewMatchers
 import android.support.test.espresso.matcher.ViewMatchers.isDisplayed
 import android.support.test.espresso.matcher.ViewMatchers.withId
 import android.support.test.espresso.matcher.ViewMatchers.withText
-import android.support.test.espresso.web.sugar.Web.onWebView
-import android.support.test.espresso.web.webdriver.DriverAtoms.findElement
-import android.support.test.espresso.web.webdriver.DriverAtoms.webClick
-import android.support.test.espresso.web.webdriver.Locator
 import android.support.test.rule.ActivityTestRule
 import android.support.v7.widget.RecyclerView
 import android.view.inputmethod.InputMethodManager
@@ -34,10 +30,8 @@ import org.hamcrest.core.Is
 import org.junit.Assert
 import org.junit.Rule
 import org.mozilla.focus.R
-import org.mozilla.focus.activity.MainActivity
 import org.mozilla.focus.activity.SettingsActivity
 import org.mozilla.focus.helper.ActivityRecreateLeakWatcherIdlingResource
-import org.mozilla.focus.helper.SessionLoadedIdlingResource
 import org.mozilla.focus.utils.FirebaseHelper
 import org.mozilla.focus.widget.TelemetrySwitchPreference
 import org.mozilla.focus.widget.TurboSwitchPreference
@@ -91,9 +85,16 @@ class ScreenshotRobot : MenuRobot() {
 class FindInPageRobot : MenuRobot() {
 
     fun findKeywordInPage(keyword: String, imeAction: Boolean = true) {
-
         onView(withId(R.id.find_in_page_query_text)).check(matches(isDisplayed())).perform(replaceText(keyword))
         if (imeAction) onView(withId(R.id.find_in_page_query_text)).perform(pressImeActionButton())
+    }
+
+    fun findKeywordInPageChecking(keyword: String) {
+        onView(withId(R.id.find_in_page_query_text)).check(matches(withText(keyword)))
+        onView(withId(R.id.find_in_page_result_text)).check(matches(isDisplayed()))
+
+        onView(withId(R.id.display_url)).check(matches(isDisplayed()))
+        onView(withId(R.id.browser_screen_menu)).check(matches(isDisplayed()))
     }
 
     fun navigateKeywordSearchInPage(forward: Boolean = false) {
@@ -111,6 +112,17 @@ class FindInPageRobot : MenuRobot() {
         onView(withId(R.id.find_in_page_prev_btn)).check(matches(isDisplayed()))
         onView(withId(R.id.find_in_page_next_btn)).check(matches(isDisplayed()))
         onView(withId(R.id.find_in_page_close_btn)).check(matches(isDisplayed()))
+    }
+
+    fun swipeOnWebview(direction: String) {
+        when (direction) {
+            "UP" -> onView(withId(R.id.webview_slot)).perform(ViewActions.swipeUp())
+            "DOWN" ->onView(withId(R.id.webview_slot)).perform(ViewActions.swipeDown())
+        }
+    }
+
+    fun findInPageClosed() {
+        onView(withId(R.id.find_in_page)).check(matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.GONE)))
     }
 }
