+++ app/src/androidTest/java/org/mozilla/focus/autobot/SessionRobot.kt
@@ -2,8 +2,10 @@ package org.mozilla.focus.autobot
 
 import android.app.Activity
 import android.content.Intent
+import android.net.Uri
 import android.support.test.InstrumentationRegistry
 import android.support.test.espresso.Espresso.onView
+import android.support.test.espresso.IdlingRegistry
 import android.support.test.espresso.action.Tap
 import android.support.test.espresso.action.ViewActions.click
 import android.support.test.espresso.action.ViewActions.pressImeActionButton
@@ -25,6 +27,7 @@ import android.support.test.uiautomator.UiSelector
 import android.view.View
 import org.hamcrest.Matchers
 import org.mozilla.focus.R
+import org.mozilla.focus.activity.FindInPageTest
 import org.mozilla.focus.activity.MainActivity
 import org.mozilla.focus.helper.ScreenshotIdlingResource
 import org.mozilla.focus.helper.SessionLoadedIdlingResource
@@ -194,12 +197,27 @@ class SessionRobot : MenuRobot() {
         activity.startActivity(intent)
     }
 
-
     fun tapLocatorOnWebView(activity: MainActivity, locatorString: String) {
         sessionLoadedIdlingResource = SessionLoadedIdlingResource(activity)
 
         onWebView()
                 .withElement(findElement(Locator.XPATH, locatorString))
                 .perform(webClick())
+    }
+
+    fun sendBrowsingIntent(activity: Activity, url: String) {
+//        sessionLoadedIdlingResource = SessionLoadedIdlingResource(activity)
+//        IdlingRegistry.getInstance().register(sessionLoadedIdlingResource)
+
+        // Simulate third party app sending browsing url intent to rocket
+        val intent = Intent()
+        intent.action = Intent.ACTION_VIEW
+        intent.data = Uri.parse(url)
+        intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+        val targetContext = InstrumentationRegistry.getInstrumentation().targetContext
+        intent.setPackage(targetContext.packageName)
+        targetContext.startActivity(intent)
+
+//        IdlingRegistry.getInstance().unregister(sessionLoadedIdlingResource)
     }
 }
